Fix dropdown position bug
Because the width of the dropdown was
automatically spanning its entire container,
right-aligning the dropdown menu was sending the
menu to the right side of the screen, even when
the dropdown toggle was on the left side of the
screen. The dropdown width has now been set to
match the width of the dropdown toggle so that the
menu is always visually tied to the button
regardless of left or right alignment.